Refactor IndexingMsg (#3590)
* fix some failing array api tests


* add support for homogenous tuple formals in 'registerCommand' annotation. Use 'registerCommand' to simplify [int] indexing message handler


* convert [pdarray] and [slice] indexing to use new registration annotations


* add support for formals with type-queried types to 'registerCommand' annotation. Convert [int]=val and [slice]=val commands to use registerCommand


* restore Commands.chpl


* add some comments to register_commands script and apply black auto-formatting


* fix handling of max_bits in indexingMsg. Add support for returning 'SymEntry' from a proc registered with 'registerCommand'.


* use sentinel value for max_bits in indexing commands if it's not set


* convert takeAlongAxis to use instantiateAndRegister


* fix failing muldidim indexing tests


* properly extract path and line numbers from ast location in register_commands.py


* fix merge conflict due to BigInt -> bigint rename


---------

+++ tests/array_api/stats_functions.py
@@ -48,7 +48,7 @@ class StatsFunctionTests(ArkoudaTest):
         a = xp.ones((10, 5, 5))
         a[0, 0, 0] = 251
 
-        self.assertEqual(xp.mean(a), 2)
+        self.assertEqual(int(xp.mean(a)), 2)
 
         a[:, 0, 0] = 26
